utils/pdf_parser.py
import pdfplumber
from docx import Document

def extract_text_from_pdf(file_path: str) -> str:
    if file_path.endswith(".docx"):
        doc = Document(file_path)
        text = "\n".join([para.text for para in doc.paragraphs])
        return text.strip()
    
    text = ""
    with pdfplumber.open(file_path) as pdf:
        for page in pdf.pages:
            text += page.extract_text() or ""
    return text.strip()