requirements.txt
fastapi==0.115.0
uvicorn==0.42.0
langgraph==1.0.10
langchain==1.2.10
langchain-groq==1.1.2
langchain-core==1.2.14
pdfplumber==0.11.9
python-multipart==0.0.22
python-dotenv==1.0.0
pydantic==2.10.0
python-docx==1.2.0
groq==0.37.1